Split unix and windows code in manifest pkg
* move arch specific const to setup
* move manifests specific const
* move template to wrapper
